+++ Assets/Script/Travaux/TravauxFinancement.cs
@@ -5,12 +5,14 @@ public class TravauxFinancement
 {
     public string id = Guid.NewGuid().ToString();
     public string description;
-    public float coutTotal;
+    // Montants en `double` — voir AchatFinancement : un `float` perd le centime
+    // au-delà d'environ 1 M€, et l'écrit déjà faux dans le JSON.
+    public double coutTotal;
     public bool emprunt;
-    public float montantEmprunte;
-    public float apportPersonnel;
-    public float tauxInteretAnnuel;
+    public double montantEmprunte;
+    public double apportPersonnel;
+    public double tauxInteretAnnuel;
     public int dureeMois;
-    public float mensualiteCalculee;
+    public double mensualiteCalculee;
     public string dateDebutTravaux;
 }
